--- a/diskio/SPI2.c
+++ b/diskio/SPI2.c
@@ -1,0 +1,3 @@
+#define SPI_PREFIX SPI2
+#define SPI_USE_MCP23S08
+#include "pic18f47q43_spi.c"