Performance tuning reduce startup time drastically.

|CP/M3.0 startup time | speed| revision | commit log |
|---------------------|------|----------|------------|
|12.3 sec|100% |2023-05-28 24dc2cf |Merge pull request #9 from hanyazou/20230505-debug-disk-write|
|11.1 sec|110% |2023-05-28 02ed421 |Reduce redundant access to the I/O expander registers|
| 9.2 sec|133% |2023-05-28 a78b443 |Change SPI clock setting of the I/O expander from 100 KHz to 2 MHz|
| 5.5 sec|223% |2023-05-28 51bf4ee |Omit CRC calculations of SD Card read command to improve performance|
| 5.5 sec|223% |2023-05-28 2e3b91a |Optimize DMA read/write (5% faster)|
| 5.2 sec|236% |2023-06-03 6fbd129 |Use console output interrupt as well as input|
| 3.8 sec|323% |2023-06-03 cdc77ef |Add PIC hardware SPI SD Card access|